♻️ Move data methods into api dir
