+++ docs/publish.sh
@@ -9,9 +9,8 @@ git clone $GIT_DOCS
 # use temp working branch for now
 cd gitbook-docs
 git checkout include-api-docs
-cd ..
 
-node migrate.js
+node ../migrate.js
 
 git add --all
 git commit -m "📒 🤖 Mintbase SDK automated doc sync"
