+++ packages/data/src/api/tokenListingCounts/tokenListingCountsByMetaId.test.ts
@@ -1,6 +1,6 @@
 import { tokenListingCountsByMetaId, TokenListingQueryResults } from './tokenListingCountsByMetaId';
 import { GraphQLClient } from 'graphql-request';
-import { GraphqlFetchingError } from '../graphql/fetch';
+import { GraphqlFetchingError } from '../../graphql/fetch';
 
 jest.mock('graphql-request');
