+++ packages/data/src/graphql/index.ts
@@ -1,3 +1,1 @@
-export * from './tokenListingCounts/tokenListingCountsByMetaId';
-export * from './ownedTokens/ownedTokens';
 export * from './codegen/index';
