+++ packages/data/src/index.ts
@@ -1,2 +1,3 @@
 export * from './constants';
 export * from './graphql/index';
+export * from './api/index';
